Move cache into getVersion
