--- app/blog/render/replaceFolderLinks/cache.js
@@ -1,25 +1,0 @@
-class Cache {
-  constructor() {
-    this.cache = new Map();
-    this.maxEntries = 10000;
-  }
-
-  set(key, value) {
-    if (this.cache.size >= this.maxEntries) {
-      const firstKey = this.cache.keys().next().value;
-      this.cache.delete(firstKey);
-    }
-    this.cache.set(key, value);
-  }
-
-  get(key) {
-    const value = this.cache.get(key);
-    if (value) {
-      this.cache.delete(key);
-      this.cache.set(key, value);
-    }
-    return value;
-  }
-}
-
-module.exports = Cache;
