+++ app/blog/render/replaceFolderLinks/getVersion.js
@@ -2,7 +2,31 @@ const config = require("config");
 const fs = require("fs-extra");
 const hash = require("helper/hash");
 const { resolve, join } = require("path");
-const Cache = require("./cache");
+
+class Cache {
+  constructor() {
+    this.cache = new Map();
+    this.maxEntries = 10000;
+  }
+
+  set(key, value) {
+    if (this.cache.size >= this.maxEntries) {
+      const firstKey = this.cache.keys().next().value;
+      this.cache.delete(firstKey);
+    }
+    this.cache.set(key, value);
+  }
+
+  get(key) {
+    const value = this.cache.get(key);
+    if (value) {
+      this.cache.delete(key);
+      this.cache.set(key, value);
+    }
+    return value;
+  }
+}
+
 const pathCache = new Cache();
 
 async function getVersion(blogID, cacheID, value) {
